+++ src/main/scala/org/advent/year2024/Day6.scala
@@ -40,8 +40,8 @@ object Day6 extends Problem[(Graph, Int)](2024, 6) {
     dfs(graph, startingGuard)
   }
 
-  override def solution1(input: (Graph, Int)): Int =
+  override def solution1(input: (Graph, Int)): Long =
     input._1().map(_.count(char => (directions.keySet + '+').contains(char) )).sum + 1
 
-  override def solution2(input: (Graph, Int)): Int = input._2
+  override def solution2(input: (Graph, Int)): Long = input._2
 }
