+++ src/main/scala/org/advent/year2024/Day8.scala
@@ -20,10 +20,10 @@ object  Day8 extends Problem[Graph](2024, 8) {
 
   override def setup(input: List[String]): Graph = Graph(input)
 
-  override def solution1(data: Graph): Int =
+  override def solution1(data: Graph): Long =
     getMatrix(data).values.flatMap(findDistances(_).toList).toSet.count(data.checkBounds)
 
-  override def solution2(data: Graph): Int =
+  override def solution2(data: Graph): Long =
     (getMatrix(data).values.flatMap(row => findDistances(row, 50).toList).toList ++
       getMatrix(data).values.flatten).toSet.count(data.checkBounds)
 }
